Close db after loading db and add GCC flag for syntax checking
- Forgot to close DB after loading

- Add a neat GCC command-line flag, -fsyntax-only. I hoped this would
reduce or eliminate the amount of command line transformation we have
to do, but unfortunately, with -fsyntax-only, we still have to remove
-o and add a "-x" parameter (gcc seems to ignore files passed on the
command line when '-' is specified to read a file from STDIN, so it
may be possible to just leave that there).  However, it should be
possible to leave all the optimization/architecture-specific flags
alone without paying the penalty for them.