Factor out common code in simpleproj-test
- Add function create-test-compile-command-json which creates a sample
compile_commands.json file in the requested directory

- Add macro with-project-and-directory which runs a specified body
with bindings for a project directory and project list.  I looked into
defining a macro that wraps ert-deftest (eg, simpleproj-deftest), but
ert-deftest unfortunately destructures arguments such as docstring &
keyword parameters itself, which meant I would have to duplicate that
destructuring to wrap the body, and then recombine it to pass the
parameters to ert-deftest, which was too dependent on internal ERT
details.